Revert "Revert "separate dakoku""
This reverts commit 310c4a73f4ba1ae11dfb76c1a1befcb8564a43d3.
